Rearranged user tab files locations in project
