+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name=".ui.home.info.FuelPricesActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.user.UserCardInfoActivity"
+            android:name=".ui.home.user.UserCardInfoActivity"
             android:exported="false" />
         <activity
             android:name=".ui.auth.AuthActivity"
@@ -27,22 +27,16 @@
             android:name=".ui.home.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.user.UserCardActivity"
+            android:name=".ui.home.user.UserCardActivity"
             android:exported="false" />
         <activity
             android:name=".ui.home.search.LocationInputActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.home.info.FuelPricesActivity"
+            android:name=".ui.home.user.UserInformationActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.user.UserInformationActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.user.UserUpdateActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.home.MainActivity"
+            android:name=".ui.home.user.UserUpdateActivity"
             android:exported="false" />
         <activity
             android:name=".SplashActivity"
