+++ app/src/main/java/com/riteh/autoshare/ui/home/user/UserCard.kt
@@ -1,4 +1,4 @@
-package com.riteh.autoshare.ui.user
+package com.riteh.autoshare.ui.home.user
 
 class UserCard(
     var cardNumber: String,
