+++ app/src/main/res/layout/activity_user_card_info.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context=".ui.user.UserCardInfoActivity">
+    tools:context=".ui.home.user.UserCardInfoActivity">
 
     <RelativeLayout
         android:id="@+id/tv_title_container"
