+++ app/src/main/res/layout/activity_user_update.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/app_white"
-    tools:context=".ui.user.UserCardActivity">
+    tools:context=".ui.home.user.UserCardActivity">
 
     <LinearLayout
         android:id="@+id/tv_title_container"
